Clean up Java 8.

+++ DEV/org.openl.j/src/org/openl/binding/impl/BusinessIntNodeBinder.java
@@ -17,7 +17,7 @@ public class BusinessIntNodeBinder extends BusinessNumberNodeBinder {
     private int getIntValue(Long number, ISyntaxNode node) throws SyntaxNodeException {
         if (number > Integer.MAX_VALUE || number < Integer.MIN_VALUE) {
             String message = String.format("Number %d is outside the valid range %d - %d",
-                number.longValue(),
+                number,
                 Integer.MIN_VALUE,
                 Integer.MAX_VALUE);
             throw SyntaxNodeExceptionUtils.createError(message, node);
